+++ packages/@tenoxui-property/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>@tenoxui/property | nousantx</title>
+    <title>Types & Properties | TenoxUI</title>
     <meta name="description" content="@tenoxui/property simple usage for development." />
     <script src="https://cdn.jsdelivr.net/npm/tenoxui@0.11"></script>
     <script src="./src/umd/full.js"></script>
