+++ packages/@tenoxui-property/src/esm/default.js
@@ -1,4 +1,4 @@
- const defaultProps = {
+export const defaultProps = {
   // padding
   p: "padding",
   pt: "paddingTop",
@@ -99,5 +99,5 @@
   shadow: "boxShadow",
   // custom property
   box: ["width", "height"],
-  "fx-parent": ["justifyContent", "alignItems"]
+  "fx-parent": ["justifyContent", "alignItems"],
 };
