+++ packages/@tenoxui-property/src/esm/full.js
@@ -1,118 +1,8 @@
-((root, factory) => {
-  if (typeof define === "function" && define.amd) {
-    // AMD (Asynchronous Module Definition)
-    define([], factory);
-  } else if (typeof exports === "object" && typeof module === "object") {
-    module.exports = factory();
-  } else if (typeof exports === "object") {
-    exports.fullProps = factory().fullProps;
-  } else {
-    root.fullProps = factory().fullProps;
-  }
-})(typeof self !== "undefined" ? self : this, () => {
-  const fullProps = {
-    // padding
-    p: "padding",
-    pt: "paddingTop",
-    pb: "paddingBottom",
-    pr: "paddingRight",
-    pl: "paddingLeft",
-    py: ["paddingTop", "paddingBottom"],
-    px: ["paddingLeft", "paddingRight"],
-    // Margin
-    m: "margin",
-    mt: "marginTop",
-    mb: "marginBottom",
-    mr: "marginRight",
-    ml: "marginLeft",
-    my: ["marginTop", "marginBottom"],
-    mx: ["marginLeft", "marginRight"],
-    // Text and font
-    fs: "fontSize",
-    fw: "fontWeight",
-    lh: "lineHeight",
-    ls: "letterSpacing",
-    ta: "textAlign",
-    c: "color",
-    td: "textDecoration",
-    ti: "textIndent",
-    tn: "textTransform",
-    tw: "textWrap",
-    ws: "wordSpacing",
-    family: "fontFamily",
-    "font-s": "fontStyle",
-    "white-space": "whiteSpace",
-    // Positioning
-    pn: "position",
-    z: "zIndex",
-    t: "top",
-    b: "bottom",
-    r: "right",
-    l: "left",
-    // Display
-    d: "display",
-    // Width and Height
-    w: "width",
-    "w-mx": "maxWidth",
-    "w-mn": "minWidth",
-    h: "height",
-    "h-mx": "maxHeight",
-    "h-mn": "minHeight",
-    // Background
-    bg: "background",
-    "bg-c": "backgroundColor",
-    "bg-i": "backgroundImage",
-    "bg-a": "backgroundAttachment",
-    "bg-or": "backgroundOrigin",
-    "bg-s": "backgroundSize",
-    "bg-r": "backgroundRepeat",
-    "bg-p": "backgroundPosition",
-    "bg-clip": "backgroundClip",
-    // Flex
-    fx: "flex",
-    "fx-d": "flexDirection",
-    "fx-w": "flexWrap",
-    "fx-b": "flexBasis",
-    "fx-g": "flexGrow",
-    "fx-s": "flexShrink",
-    // Gap
-    g: "gap",
-    "g-y": "columnGap",
-    "g-x": "rowGap",
-    // Align
-    "a-i": "alignItems",
-    // Justify
-    "j-c": "justifyContent",
-    // Border
-    border: "border",
-    bc: "borderColor",
-    br: "borderRadius",
-    bs: "borderStyle",
-    bw: "borderWidth",
-    "bw-l": "borderLeftWidth",
-    "bw-r": "borderRightWidth",
-    "bw-t": "borderTopWidth",
-    "bw-b": "borderBottomWidth",
-    // Cursor
-    cursor: "cursor",
-    // Overflow
-    over: "overflow",
-    "over-y": "overflowY",
-    "over-x": "overflowX",
-    // Aspect Ratio
-    ratio: "aspectRatio",
-    // Transition
-    transition: "transition",
-    "tr-time": "transitionDuration",
-    "tr-prop": "transitionProperty",
-    "tr-timing": "transitionTimingFunction",
-    "tr-delay": "transitionDelay",
-    // More
-    shadow: "boxShadow",
-    // custom property
-    box: ["width", "height"],
-    "fx-parent": ["justifyContent", "alignItems"],
+import { defaultProps } from "./default.js";
 
+export const fullProps = {
+  ...defaultProps,
+  ...{
     order: "order",
     "a-c": "alignContent",
     "a-s": "alignSelf",
@@ -155,7 +45,7 @@ translate(var(--tx_translate, 0))
 translateY(var(--tx_translate-y, 0))
 translateX(var(--tx_translate-x, 0))
 translateZ(var(--tx_translate-z, 0))
-  `
+  `,
     },
     rotate: "--tx_rotate",
     "rotate-y": "--tx_rotate-y",
@@ -186,7 +76,7 @@ opacity(var(--tx_opacity, 1))
 saturate(var(--tx_saturate, 1)) 
 sepia(var(--tx_sepia, 0))
 drop-shadow(var(--tx_drop-shadow, 0 0 0 rgb(0 0 0 / 0)))
-  `
+  `,
     },
     blur: "--tx_blur",
     brightness: "--tx_brightness",
@@ -210,7 +100,7 @@ invert(var(--back_invert, 0))
 opacity(var(--back_opacity, 1)) 
 saturate(var(--back_saturate, 1)) 
 sepia(var(--back_sepia, 0))
-  `
+  `,
     },
     "back-blur": "--back_blur",
     "back-brightness": "--back_brightness",
@@ -220,7 +110,6 @@ sepia(var(--back_sepia, 0))
     "back-invert": "--back_invert",
     "back-opacity": "--back_opacity",
     "back-saturate": "--back_saturate",
-    "back-sepia": "--back_sepia"
-  };
-  return {fullProps}
-});
+    "back-sepia": "--back_sepia",
+  },
+};
