+++ packages/tenoxui/package.json
@@ -2,6 +2,18 @@
   "name": "tenoxui",
   "version": "0.11.4",
   "description": "A CSS Framework without CSS file :D",
+  "author": "NOuSantx <nousantx@gmail.com>",
+  "license": "MIT",
+  "homepage": "https://tenoxui.web.app",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/tenoxui/tenoxui.git",
+    "directory": "packages/tenoxui"
+  },
+  "bugs": {
+    "url": "https://github.com/tenoxui/tenoxui/issues"
+  },
+  "keywords": ["framework", "css", "css-in-js", "utility-first"],
   "main": "./dist/js/tenoxui.js",
   "style": "./dist/css/tenoxui.css",
   "module": "./dist/js/tenoxui.esm.js",
@@ -9,24 +21,13 @@
   "files": ["dist/*", "package.json", "README.md", "LICENSE"],
   "scripts": {
     "dev": "tsc --watch",
-    "build-css": "cleancss src/css/tenoxui.css -o dist/css/tenoxui.min.css --source-map && cp src/css/tenoxui.css dist/css/",
-    "build-js": "uglifyjs ./src/js/tenoxui.js -c -m toplevel  -o dist/js/tenoxui.min.js --source-map \"url='tenoxui.min.js.map'\" --comments \"/^!/\" --keep-fnames && uglifyjs src/js/tenoxui.js --beautify -o dist/js/tenoxui.js --source-map \"url='tenoxui.js.map'\" --comments \"/^!/\" && cp src/js/tenoxui.d.ts src/js/tenoxui.js.map dist/js && prettier -w dist/js/tenoxui.js dist/js/tenoxui.d.ts",
-    "build-esm": "uglifyjs ./src/js/tenoxui.esm.js -c -m toplevel -o dist/js/tenoxui.esm.min.js --source-map \"url='tenoxui.esm.min.js.map'\" --comments \"/^!/\" --keep-fnames && uglifyjs ./src/js/tenoxui.esm.js --beautify -o dist/js/tenoxui.esm.js --source-map \"url='tenoxui.esm.js.map'\" --comments \"/^!/\" && cp src/js/tenoxui.esm.d.ts  src/js/tenoxui.esm.js.map dist/js && prettier -w dist/js/tenoxui.esm.js dist/js/tenoxui.esm.d.ts",
-    "build": "tsc && npm run build-js && npm run build-esm && npm run build-css && echo '\\033[92m Have Fun :D'",
-    "test": "echo 'Done!'",
+    "build:css": "cleancss src/css/tenoxui.css -o dist/css/tenoxui.min.css --source-map && cp src/css/tenoxui.css dist/css/",
+    "build:js": "uglifyjs ./src/js/tenoxui.js -c -m toplevel  -o dist/js/tenoxui.min.js --source-map \"url='tenoxui.min.js.map'\" --comments \"/^!/\" --keep-fnames && uglifyjs src/js/tenoxui.js --beautify -o dist/js/tenoxui.js --source-map \"url='tenoxui.js.map'\" --comments \"/^!/\" && cp src/js/tenoxui.d.ts src/js/tenoxui.js.map dist/js && prettier -w dist/js/tenoxui.js dist/js/tenoxui.d.ts",
+    "build:esm": "uglifyjs ./src/js/tenoxui.esm.js -c -m toplevel -o dist/js/tenoxui.esm.min.js --source-map \"url='tenoxui.esm.min.js.map'\" --comments \"/^!/\" --keep-fnames && uglifyjs ./src/js/tenoxui.esm.js --beautify -o dist/js/tenoxui.esm.js --source-map \"url='tenoxui.esm.js.map'\" --comments \"/^!/\" && cp src/js/tenoxui.esm.d.ts src/js/tenoxui.esm.js.map dist/js && prettier -w dist/js/tenoxui.esm.js dist/js/tenoxui.esm.d.ts",
+    "build": "tsc && yarn build:css && yarn build:js && yarn build:esm && echo '\\033[92mHave Fun :D\\033[0m'",
+    "test": "echo '\\033[92mHave Fun :D\\033[0m'",
     "prepublish": "npm pack --dry-run"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/tenoxui/css.git"
-  },
-  "keywords": ["framework", "css", "css-in-js", "utility-first"],
-  "author": "NOuSantx",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/tenoxui/css/issues"
-  },
-  "homepage": "https://tenoxui.web.app",
   "devDependencies": {
     "clean-css-cli": "^5.6.3",
     "prettier": "^3.2.5",
