--- packages/tenoxui/src/lib/property.esm.js
@@ -1,124 +1,0 @@
-/*!
- * @tenoxui/property v1.1.0 (https://github.com/tenoxui/css)
- * Copyright (c) 2024 NOuSantx
- * Licensed under the MIT License (https://github.com/tenoxui/css/blob/main/LICENSE)
- */
-export const property = {
-  p: "padding",
-  pt: "paddingTop",
-  pb: "paddingBottom",
-  pr: "paddingRight",
-  pl: "paddingLeft",
-  py: ["paddingTop", "paddingBottom"],
-  px: ["paddingLeft", "paddingRight"],
-  m: "margin",
-  mt: "marginTop",
-  mb: "marginBottom",
-  mr: "marginRight",
-  ml: "marginLeft",
-  my: ["marginTop", "marginBottom"],
-  mx: ["marginLeft", "marginRight"],
-  fs: "fontSize",
-  fw: "fontWeight",
-  lh: "lineHeight",
-  ls: "letterSpacing",
-  ta: "textAlign",
-  tc: "color",
-  td: "textDecoration",
-  ti: "textIndent",
-  tn: "textTransform",
-  ws: "wordSpacing",
-  family: "fontFamily",
-  "font-s": "fontStyle",
-  "white-space": "whiteSpace",
-  position: "position",
-  z: "zIndex",
-  t: "top",
-  b: "bottom",
-  r: "right",
-  l: "left",
-  d: "display",
-  w: "width",
-  "w-mx": "maxWidth",
-  "w-mn": "minWidth",
-  h: "height",
-  "h-mx": "maxHeight",
-  "h-mn": "minHeight",
-  bg: "background",
-  "bg-color": "backgroundColor",
-  "bg-image": "backgroundImage",
-  "bg-attach": "backgroundAttachment",
-  "bg-origin": "backgroundOrigin",
-  "bg-size": "backgroundSize",
-  "bg-clip": "backgroundClip",
-  "bg-repeat": "backgroundRepeat",
-  "bg-loc": "backgroundPosition",
-  flex: "flex",
-  "flex-auto": "flex",
-  fd: "flexDirection",
-  "flex-wrap": "flexWrap",
-  "flex-basis": "flexBasis",
-  "flex-grow": "flexGrow",
-  "flex-shrink": "flexShrink",
-  order: "order",
-  gap: "gap",
-  ac: "alignContent",
-  ai: "align-items",
-  jc: "justifyContent",
-  ji: "justifyItems",
-  filter: "filter",
-  blur: "ftr",
-  brightness: "ftr",
-  contrast: "ftr",
-  grayscale: "ftr",
-  "hue-rotate": "ftr",
-  saturate: "ftr",
-  sepia: "ftr",
-  "drop-shadow": "drop-shadow",
-  opa: "opacity",
-  backdrop: "backdropFilter",
-  "back-blur": "bFt",
-  "back-brightness": "bFt",
-  "back-contrast": "bFt",
-  "back-grayscale": "bFt",
-  "back-saturate": "bFt",
-  "back-sepia": "bFt",
-  br: "borderRadius",
-  bw: "borderWidth",
-  "bw-left": "borderLeftWidth",
-  "bw-right": "borderRightWidth",
-  "bw-top": "borderTopWidth",
-  "bw-bottom": "borderBottomWidth",
-  bs: "borderStyle",
-  "radius-top": ["borderTopLeftRadius", "borderTopRightRadius"],
-  "radius-bottom": ["borderBottomLeftRadius", "borderBottomRightRadius"],
-  "radius-left": ["borderTopLeftRadius", "borderBottomLeftRadius"],
-  "radius-right": ["borderTopRightRadius", "borderBottomRightRadius"],
-  cursor: "cursor",
-  over: "overflow",
-  "over-y": "overflowY",
-  "over-x": "overflowX",
-  ratio: "aspectRatio",
-  transition: "transition",
-  "tr-time": "transitionDuration",
-  "tr-prop": "transitionProperty",
-  "tr-timing": "transitionTimingFunction",
-  "tr-delay": "transitionDelay",
-  transform: "transform",
-  "move-x": "tra",
-  "move-y": "tra",
-  matrix: "tra",
-  "matrix-3d": "tra",
-  "rt-3d": "tra",
-  translate: "tra",
-  "scale-3d": "tra",
-  "scale-x": "tra",
-  "scale-y": "tra",
-  skew: "tra",
-  "skew-x": "tra",
-  "skew-y": "tra",
-  rt: "rotate",
-  scale: "scale",
-  box: ["width", "height"],
-  "flex-parent": ["justifyContent", "alignItems"]
-};
