+++ utils/generate-color.mjs
